pyproject.toml
[tool.poetry]
name = "llm-chatbot"
version = "0.1.0"
description = ""
authors = ["Philip Adzanoukpe <epigos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
langchain = "^0.3.13"
langchain-openai = "^0.2.14"
langchain-community = "^0.3.13"
faiss-cpu = "^1.9.0.post1"
chainlit = "^1.3.2"
structlog = "^24.4.0"
pydantic-settings = "^2.7.1"
pydantic = "2.10.1"
langgraph = "^0.2.60"
mem0ai = "^0.1.40"
langchain-anthropic = "^0.3.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"
isort = "^5.13.2"
black = "^24.10.0"
autoflake = "^2.3.1"
pre-commit = "^4.0.1"
ipython = "^8.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
known_first_party = "tests"
line_length = 88
multi_line_output = 3
use_parentheses = true
skip_glob = ".venv/*,.aws-sam/*,migrations/*"

[tool.black]
target-version = ['py311']

[tool.mypy]
strict = true
exclude = ["tests", "dist", "migrations"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "main.*"
ignore_errors = true

[tool.autoflake]
quiet = true
remove-all-unused-imports = true
recursive = true
